+++ localstack/aws/handlers/auth.py
@@ -1,8 +1,8 @@
 import logging
 
+from localstack.aws.accounts import get_account_id_from_access_key_id, set_ctx_aws_access_key_id
 from localstack.constants import HEADER_LOCALSTACK_ACCOUNT_ID
 from localstack.http import Response
-from localstack.utils.accounts import get_account_id_from_access_key_id, set_ctx_aws_access_key_id
 from localstack.utils.aws.aws_stack import extract_access_key_id_from_auth_header
 
 from ..api import RequestContext
