+++ localstack/services/awslambda/event_source_listeners/kinesis_event_source_listener.py
@@ -1,11 +1,11 @@
 import base64
 from typing import Dict, List, Optional
 
+from localstack.aws.accounts import get_aws_account_id
 from localstack.services.awslambda.event_source_listeners.stream_event_source_listener import (
     StreamEventSourceListener,
 )
 from localstack.services.awslambda.lambda_api import get_event_sources
-from localstack.utils.accounts import get_aws_account_id
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import first_char_to_lower, to_str
 from localstack.utils.threads import FuncThread
