+++ localstack/services/logs/provider.py
@@ -13,6 +13,7 @@ from moto.logs.models import LogsBackend
 from moto.logs.models import LogStream as MotoLogStream
 from moto.logs.models import logs_backends
 
+from localstack.aws.accounts import get_aws_account_id
 from localstack.aws.api import RequestContext
 from localstack.aws.api.logs import (
     InputLogEvents,
@@ -27,7 +28,6 @@ from localstack.constants import APPLICATION_AMZ_JSON_1_1
 from localstack.services.messages import Request, Response
 from localstack.services.moto import MotoFallbackDispatcher, call_moto
 from localstack.services.plugins import ServiceLifecycleHook
-from localstack.utils.accounts import get_aws_account_id
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import is_number
 from localstack.utils.patch import patch
