+++ localstack/services/sqs/legacy/sqs_listener.py
@@ -9,11 +9,11 @@ from moto.sqs.utils import parse_message_attributes
 from requests.models import Request
 
 from localstack import config
+from localstack.aws.accounts import get_aws_account_id
 from localstack.config import SQS_PORT_EXTERNAL
 from localstack.services.install import SQS_BACKEND_IMPL
 from localstack.services.sns.provider import unsubscribe_sqs_queue
 from localstack.services.sqs.utils import is_sqs_queue_url
-from localstack.utils.accounts import get_aws_account_id
 from localstack.utils.analytics import event_publisher
 from localstack.utils.aws import aws_stack
 from localstack.utils.aws.aws_responses import (
