+++ tests/integration/secretsmanager/test_secretsmanager.py
@@ -8,9 +8,9 @@ import pytest
 import requests
 from py._code.code import ExceptionInfo
 
+from localstack.aws.accounts import get_aws_account_id
 from localstack.services.awslambda.lambda_utils import LAMBDA_RUNTIME_PYTHON36
 from localstack.utils import testutil
-from localstack.utils.accounts import get_aws_account_id
 from localstack.utils.aws import aws_stack
 from localstack.utils.strings import short_uid
 from localstack.utils.time import today_no_time
