+++ tests/unit/services/opensearch/test_cluster_manager.py
@@ -1,9 +1,9 @@
 import pytest
 
 from localstack import config
+from localstack.aws.accounts import get_aws_account_id
 from localstack.aws.api.opensearch import EngineType
 from localstack.services.opensearch.cluster_manager import DomainKey, build_cluster_endpoint
-from localstack.utils.accounts import get_aws_account_id
 
 
 class TestBuildClusterEndpoint:
